src/test/java/uk/gov/hmcts/divorce/systemupdate/event/SystemMigrateBulkCaseTest.java
package uk.gov.hmcts.divorce.systemupdate.event;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
import uk.gov.hmcts.ccd.sdk.api.Event;
import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
import uk.gov.hmcts.divorce.bulkaction.ccd.BulkActionState;
import uk.gov.hmcts.divorce.bulkaction.data.BulkActionCaseData;
import uk.gov.hmcts.divorce.divorcecase.model.UserRole;

import static org.assertj.core.api.Assertions.assertThat;
import static uk.gov.hmcts.divorce.systemupdate.event.SystemMigrateBulkCase.SYSTEM_MIGRATE_BULK_CASE;
import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.createBulkActionConfigBuilder;
import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.getEventsFrom;

@ExtendWith(SpringExtension.class)
public class SystemMigrateBulkCaseTest {

    @InjectMocks
    private SystemMigrateBulkCase systemMigrateBulkCase;

    @Test
    void shouldAddConfigurationToConfigBuilder() {
        final ConfigBuilderImpl<BulkActionCaseData, BulkActionState, UserRole> configBuilder = createBulkActionConfigBuilder();

        systemMigrateBulkCase.configure(configBuilder);

        assertThat(getEventsFrom(configBuilder).values())
            .extracting(Event::getId)
            .contains(SYSTEM_MIGRATE_BULK_CASE);
    }

    @Test
    void shouldReturnCase() {
        final CaseDetails<BulkActionCaseData, BulkActionState> details = new CaseDetails<>();
        final BulkActionCaseData caseData = BulkActionCaseData.builder().build();
        details.setData(caseData);
        details.setId(1L);

        final AboutToStartOrSubmitResponse<BulkActionCaseData, BulkActionState> response =
            systemMigrateBulkCase.aboutToSubmit(details, details);

        assertThat(response.getData()).isEqualTo(caseData);
    }
}
